Inhibit drift learning while there is backlash potential
If backlash calibration measures a large backlash but the gears
move early, that only means movement is possible, but may still
mean the axis doesn't yet move at full speed.

If the backlash flag is cleared as soon as movement is detected,
drift compensation will mistake a slow-moving gear for actual
drift.

Slow-moving gear is another type of error, that requires
multiplicative compensation rather than additive compensation
which is what drift learning does. So until we get the multiplicative
part done, learning has to be stopped while the gears are
in a compromised state.