Ensure filter position reproducibility
CFW may not land on the same exact position when changing to
a filter from different directions.

Make sure each filter is approach in exactly the same manner
at all times, by doing an initial cycle through the filter
sequence before starting captures. This ensures the first
filter is approached from the same direction as it would when
starting a new loop.

Such reproducibility is important for proper flat calibration