Prevent "contamined" drift measurements during guiding
Make sure to only measure drift during stable portions of
guiding. Exclude samples that had ignored pulses to prevent
backlash, or samples that had large pulses, as those tend to
make it harder to predict the response of the mount to the guide
pulses and thus harder to separate pulse effects from actual drift.

Stop reconstructing guide pulse on drift updates. That strategy
only worked when drift guiding was the main guiding method, but
we're fully embracing pulses now so that doesn't make sense anymore