Separate RA/DEC target pulse settings
RA tends to better handle shorter pulses even if the drive
itself is inaccurate because there's a background "medium"
of pulses for sideral tracking that can be perturbed rather
precisely.

DEC on the other hand depends fully on the precision of the
motor drive, so it makes sense to have separate settings
for them.