Don't leave sub-minimum pulses in the accumulator
If an accumulated pulse is just slightly above the maximum/target
during a cycle, a small residue will be left in the accumulator
that may induce noise in following cycles, especially since small
pulses are not accurately executed in most mounts.

So prefer a slightly longer than usual pulse instead of an invisible
tiny leftover in the accumulators.