Optimize image load/save
Add clipping arithmetic to encode/decode_srgb so that a direct
convertion to/from encoded ranges can be done, instead of doing
the conversion in steps.

Exploit this in image loading/save methods that have to deal
with nonlinear encodings to avoid quite a bit of temporaries.

When possible, try to stick to LUT-based conversions as well.