Work around a bug in rawpy and other small improvements
Works around a bug in which rawpy starts failing to load
RAW metadata after a number of images have been processed.

This workaround makes the code depend less on the presence
of the img.rimg.raw_pattern attribute, and instead uses the
cached one from BaseRop which does not suffer that bug.

This is only effective as long as rawpy still is able to
read the raw data itself, which it seems in some versions at
least is the case.

Other small improvements intermingled in this change:

- InvFWHM overrides literal parsing so the CLI can take more
  natural thresholds when applying min/max rank filtering
  during subframe selection
- Added stretch:black ROP that only moves the black point